--- a/src/libraries/System.Private.CoreLib/src/System/DateTime.Windows.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/DateTime.Windows.cs
@@ -1,48 +1,13 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace System
 {
     public readonly partial struct DateTime
     {
         internal static readonly bool s_systemSupportsLeapSeconds = SystemSupportsLeapSeconds();
-
-        public static unsafe DateTime UtcNow
-        {
-            get
-            {
-                ulong fileTime;
-                s_pfnGetSystemTimeAsFileTime(&fileTime);
-
-                if (s_systemSupportsLeapSeconds)
-                {
-                    Interop.Kernel32.SYSTEMTIME time;
-                    ulong hundredNanoSecond;
-
-                    if (Interop.Kernel32.FileTimeToSystemTime(&fileTime, &time) != Interop.BOOL.FALSE)
-                    {
-                        // to keep the time precision
-                        ulong tmp = fileTime; // temp. variable avoids double read from memory
-                        hundredNanoSecond = tmp % TicksPerMillisecond;
-                    }
-                    else
-                    {
-                        Interop.Kernel32.GetSystemTime(&time);
-                        hundredNanoSecond = 0;
-                    }
-
-                    return CreateDateTimeFromSystemTime(in time, hundredNanoSecond);
-                }
-                else
-                {
-                    return new DateTime(fileTime + FileTimeOffset | KindUtc);
-                }
-            }
-        }
 
         internal static unsafe bool IsValidTimeWithLeapSeconds(int year, int month, int day, int hour, int minute, DateTimeKind kind)
         {
@@ -80,7 +45,9 @@ namespace System
             {
                 throw new ArgumentOutOfRangeException(nameof(fileTime), SR.ArgumentOutOfRange_DateTimeBadTicks);
             }
-            return CreateDateTimeFromSystemTime(in time, fileTime % TicksPerMillisecond);
+
+            ulong ticks = GetTicksFromSystemTime(in time, fileTime % TicksPerMillisecond);
+            return new DateTime(ticks | KindUtc);
         }
 
         private static unsafe ulong ToFileTimeLeapSecondsAware(long ticks)
@@ -110,7 +77,7 @@ namespace System
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static DateTime CreateDateTimeFromSystemTime(in Interop.Kernel32.SYSTEMTIME time, ulong hundredNanoSecond)
+        internal static ulong GetTicksFromSystemTime(in Interop.Kernel32.SYSTEMTIME time, ulong hundredNanoSecond)
         {
             uint year = time.Year;
             uint[] days = IsLeapYear((int)year) ? s_daysToMonth366 : s_daysToMonth365;
@@ -124,52 +91,12 @@ namespace System
             if (second <= 59)
             {
                 ulong tmp = second * (uint)TicksPerSecond + time.Milliseconds * (uint)TicksPerMillisecond + hundredNanoSecond;
-                return new DateTime(ticks + tmp | KindUtc);
+                return ticks + tmp;
             }
 
             // we have a leap second, force it to last second in the minute as DateTime doesn't account for leap seconds in its calculation.
             // we use the maxvalue from the milliseconds and the 100-nano seconds to avoid reporting two out of order 59 seconds
-            ticks += TicksPerMinute - 1 | KindUtc;
-            return new DateTime(ticks);
-        }
-
-        private static unsafe readonly delegate* unmanaged[SuppressGCTransition]<ulong*, void> s_pfnGetSystemTimeAsFileTime = GetGetSystemTimeAsFileTimeFnPtr();
-
-        private static unsafe delegate* unmanaged[SuppressGCTransition]<ulong*, void> GetGetSystemTimeAsFileTimeFnPtr()
-        {
-            IntPtr kernel32Lib = Interop.Kernel32.LoadLibraryEx(Interop.Libraries.Kernel32, IntPtr.Zero, Interop.Kernel32.LOAD_LIBRARY_SEARCH_SYSTEM32);
-            Debug.Assert(kernel32Lib != IntPtr.Zero);
-
-            IntPtr pfnGetSystemTime = NativeLibrary.GetExport(kernel32Lib, "GetSystemTimeAsFileTime");
-
-            if (NativeLibrary.TryGetExport(kernel32Lib, "GetSystemTimePreciseAsFileTime", out IntPtr pfnGetSystemTimePrecise))
-            {
-                // GetSystemTimePreciseAsFileTime exists and we'd like to use it.  However, on
-                // misconfigured systems, it's possible for the "precise" time to be inaccurate:
-                //     https://github.com/dotnet/runtime/issues/9014
-                // If it's inaccurate, though, we expect it to be wildly inaccurate, so as a
-                // workaround/heuristic, we get both the "normal" and "precise" times, and as
-                // long as they're close, we use the precise one. This workaround can be removed
-                // when we better understand what's causing the drift and the issue is no longer
-                // a problem or can be better worked around on all targeted OSes.
-
-                // Retry this check several times to reduce chance of false negatives due to thread being rescheduled
-                // at wrong time.
-                for (int i = 0; i < 10; i++)
-                {
-                    long systemTimeResult, preciseSystemTimeResult;
-                    ((delegate* unmanaged[SuppressGCTransition]<long*, void>)pfnGetSystemTime)(&systemTimeResult);
-                    ((delegate* unmanaged[SuppressGCTransition]<long*, void>)pfnGetSystemTimePrecise)(&preciseSystemTimeResult);
-
-                    if (Math.Abs(preciseSystemTimeResult - systemTimeResult) <= 100 * TicksPerMillisecond)
-                    {
-                        pfnGetSystemTime = pfnGetSystemTimePrecise; // use the precise version
-                        break;
-                    }
-                }
-            }
-
-            return (delegate* unmanaged[SuppressGCTransition]<ulong*, void>)pfnGetSystemTime;
+            return ticks + TicksPerMinute - 1;
         }
     }
 }