--- a/src/libraries/System.Private.CoreLib/src/System/Diagnostics/ActualSystemClock.Unix.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Diagnostics/ActualSystemClock.Unix.cs
@@ -1,0 +1,10 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace System.Diagnostics
+{
+    public partial class ActualSystemClock
+    {
+        private static unsafe ulong GetTicks() => (ulong)(Interop.Sys.GetSystemTimeAsTicks() + DateTime.UnixEpochTicks);
+    }
+}