This PR is a proposed implementation of the [revised API proposal I submitted](https://github.com/dotnet/runtime/issues/36617#issuecomment-784734677) for #36617.  **I do not expect this PR to be merged at this time.** Please see the issue thread for details.

One open question I have with specific regards to the implementation (not the public API)

- In `TimeContext.cs`, I implement the internal storage for the ambient context simply as:

    ```csharp
    private static readonly AsyncLocal<TimeContext> s_context = new();
    ```

  This seems to work fine, and all tests pass.  However, looking at [the implementation](https://github.com/dotnet/runtime/blob/ecfcdfd776da13f12f521e3b5b9a13483a0a5038/src/libraries/System.Private.CoreLib/src/System/Globalization/CultureInfo.cs#L116-L132) for a similar concept, I see that `CultureInfo.CurrentCulture` and `CultureInfo.CurrentUICulture` each use *both* an `AsyncLocal<T>` field and a `[ThreadStatic]` field, with some logic synchronizing them.  Is that necessary here also? Why or why not?

Thanks.

# Benchmarks

I ran the existing microbenchmarks [from here](https://github.com/dotnet/performance/blob/master/src/benchmarks/micro/libraries/System.Runtime), that test `DateTime.UtcNow`, `DateTimeOffset.UtcNow`, `DateTime.Now`, and `DateTimeOffset.Now`, both before and after my changes, and both with and without leap seconds enabled (via registry flag [as described here](https://github.com/dotnet/runtime/issues/13091#issuecomment-511937942)).  The benchmark results show roughly a 8-20ns cost for the `UtcNow` properties, and a 19-92ns cost for the `Now` properties, both of which I believe are reasonable and negligible.

Full benchmark results follow.

## ResultsComparer difference, with leap seconds enabled

| Slower                                     | diff/base | Base Median (ns) | Diff Median (ns) | Modality |
| ------------------------------------------ | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_DateTime.GetUtcNow       |      1.23 |            84.10 |           103.49 | bimodal  |
| System.Tests.Perf_DateTimeOffset.GetUtcNow |      1.14 |            87.15 |            99.20 |          |
| System.Tests.Perf_DateTime.GetNow          |      1.41 |           225.26 |           317.44 | bimodal  |
| System.Tests.Perf_DateTimeOffset.GetNow    |      1.09 |           266.76 |           291.16 |          |

## ResultsComparer difference, with leap seconds disabled

| Slower                                     | diff/base | Base Median (ns) | Diff Median (ns) | Modality |
| ------------------------------------------ | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_DateTime.GetUtcNow       |      1.29 |            29.12 |            37.46 |          |
| System.Tests.Perf_DateTimeOffset.GetUtcNow |      1.44 |            33.19 |            47.91 |          |
| System.Tests.Perf_DateTime.GetNow          |      1.37 |           172.73 |           236.09 | several? |
| System.Tests.Perf_DateTimeOffset.GetNow    |      1.09 |           214.23 |           232.70 |          |

## Test Setup

``` ini
BenchmarkDotNet=v0.12.1.1466-nightly, OS=Windows 10.0.19042
Intel Core i7-6600U CPU 2.60GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK=6.0.100-preview.2.21120.3
  [Host]     : .NET 6.0.0 (6.0.21.11806), X64 RyuJIT
  Job-YNAJFG : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  Toolchain=CoreRun  
IterationTime=250.0000 ms  MaxIterationCount=20  MinIterationCount=15  
WarmupCount=1  
```

## Before changes, leap seconds enabled

**Benchmark Test Class: `System.Tests.Perf_DateTime`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|---------:|------:|------:|------:|----------:|
|    GetNow | 224.61 ns | 9.719 ns | 10.80 ns | 225.26 ns | 200.93 ns | 248.0 ns |     - |     - |     - |         - |
| GetUtcNow |  88.22 ns | 9.425 ns | 10.85 ns |  84.10 ns |  75.37 ns | 108.9 ns |     - |     - |     - |         - |

**Benchmark Test Class: `System.Tests.Perf_DateTimeOffset`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|    GetNow | 264.16 ns | 7.296 ns | 8.403 ns | 266.76 ns | 248.45 ns | 276.40 ns |     - |     - |     - |         - |
| GetUtcNow |  87.16 ns | 0.460 ns | 0.384 ns |  87.15 ns |  86.67 ns |  88.07 ns |     - |     - |     - |         - |

## After changes, leap seconds enabled

**Benchmark Test Class: `System.Tests.Perf_DateTime`**

|    Method |     Mean |    Error |   StdDev |   Median |       Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|---------:|---------:|---------:|----------:|---------:|-------:|------:|------:|----------:|
|    GetNow | 318.6 ns | 17.70 ns | 19.67 ns | 317.4 ns | 295.43 ns | 371.8 ns | 0.0300 |     - |     - |      64 B |
| GetUtcNow | 102.3 ns | 12.35 ns | 14.22 ns | 103.5 ns |  81.61 ns | 126.0 ns |      - |     - |     - |         - |

**Benchmark Test Class: `System.Tests.Perf_DateTimeOffset`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|---------:|------:|------:|------:|----------:|
|    GetNow | 290.33 ns | 7.544 ns | 8.072 ns | 291.16 ns | 274.91 ns | 309.7 ns |     - |     - |     - |         - |
| GetUtcNow |  98.69 ns | 2.394 ns | 2.757 ns |  99.20 ns |  93.35 ns | 102.8 ns |     - |     - |     - |         - |

## Before changes, leap seconds disabled

**Benchmark Test Class: `System.Tests.Perf_DateTime`**

|    Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|    GetNow | 181.71 ns | 18.316 ns | 21.093 ns | 172.73 ns | 155.07 ns | 228.21 ns |     - |     - |     - |         - |
| GetUtcNow |  28.67 ns |  0.906 ns |  0.930 ns |  29.12 ns |  26.47 ns |  29.91 ns |     - |     - |     - |         - |

**Benchmark Test Class: `System.Tests.Perf_DateTimeOffset`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|    GetNow | 215.47 ns | 3.810 ns | 4.235 ns | 214.23 ns | 208.70 ns | 224.72 ns |     - |     - |     - |         - |
| GetUtcNow |  32.98 ns | 1.020 ns | 1.092 ns |  33.19 ns |  30.72 ns |  34.82 ns |     - |     - |     - |         - |

## After changes, leap seconds disabled

**Benchmark Test Class: System.Tests.`Perf_DateTime`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |       Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
|    GetNow | 235.03 ns | 8.502 ns | 8.731 ns | 236.09 ns | 218.86 ns | 255.13 ns | 0.0302 |     - |     - |      64 B |
| GetUtcNow |  37.37 ns | 1.473 ns | 1.696 ns |  37.46 ns |  34.88 ns |  40.53 ns |      - |     - |     - |         - |

**Benchmark Test Class: `System.Tests.Perf_DateTimeOffset`**

|    Method |      Mean |    Error |   StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|    GetNow | 231.58 ns | 8.161 ns | 8.380 ns | 232.70 ns | 214.52 ns | 247.20 ns |     - |     - |     - |         - |
| GetUtcNow |  47.59 ns | 1.233 ns | 1.319 ns |  47.91 ns |  43.86 ns |  49.18 ns |     - |     - |     - |         - |