@davidfowl 2021-02-24
The proposed API needs to go through a couple of phases before it gets to the PR stage.