@tarekgh 2021-02-24
@mattjohnsonpint in https://github.com/dotnet/runtime/issues/36617 it was discuss that we don't want to allow mocking APIs in the core and that should be supported in the extensions if we need to after getting of duplicates. So I don't think we'll accept such changes before finalizing the discussion. 

I am adding here the folks who attended the design review for this issue: @GrabYourPitchforks @terrajobst @bartonjs