@mattjohnsonpint 2021-02-24
Agreed about needing to go through more discussion and phases.  I provide the PR so you could see how such an API might be implemented.  We'll continue accordingly.

On the point of mocking, please see [my comments in the issue thread](https://github.com/dotnet/runtime/issues/36617#issuecomment-784734677).  Thanks.