@tarekgh 2021-02-24
> The benchmark results show roughly a 8-20ns cost for the UtcNow properties, and a 19-92ns cost for the Now properties, both of which I believe are reasonable and negligible.

If I am reading the benchmark correctly, it will be non-trivial regression. we have been working to enhance the perf of the Now/UtcNow trying to get every gain as possible. The regression percentage is not trivial. don't look at the time units but look at the percentage and consider the scenarios that the app is calling Now/UtcNow more frequent. We are trying to think in ways to enhance the performance and such change would make that difficult.