@davidfowl 2021-02-24
I think this API is important but I'm not a fan of the async local usage