@mattjohnsonpint 2021-02-24
@tarekgh, I saw several things that could improve perf while I was working on this.  Some are in here, but several were unrelated to this feature.  I can send a separate PR for those things if you like.

Generally agree about the perf.  Faster is always better.  Not sure that a few nanoseconds difference is really all that much though.  I did find it odd that the benchmarks showed the `DateTimeOffset.UtcNow` test faster (by percent) with leap seconds *enabled*, given that's the slower path.  I seemed to get quite a bit of variation in the benchmarks, even though my system was as quiet as possible at the time.  Perhaps they don't do well at such small scales?  Not sure.