@mattjohnsonpint 2021-02-24
Would love to find a better way than `AsyncLocal` also.  I couldn't think of a better way that could affect the results from existing APIs.  Seemed like the right approach, since it's used by things like `CultureInfo.CurrentCulture`.  I'm open to suggestions.